Collegespace Landing Page
This Landing Page was created by me and Rachita Gupta for our society
Collegespace to serve as the landing page of the website.

It was up from July 2014 up till March 2015, when it was made from
scratch again by a larger group of developers in the society.

But the theme of this webpage became so illustrious that its screenshot
is used in the new webpage.